# docs/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台管理接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/user/epay/notify": {
      "get": {
        "summary": "易支付回调",
        "deprecated": false,
        "description": "🔓 无需鉴权（支付回调）",
        "tags": [
          "充值"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "SessionAuth1": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken1": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 Qing-Api-User"
      },
      "QingApiUser1": {
        "type": "apiKey",
        "in": "header",
        "name": "Qing-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qing_api_refresh",
        "description": "仅 refresh/logout 使用的 HttpOnly、SameSite=Strict Cookie"
      },
      "SecurityProof": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Security-Proof",
        "description": "绑定用户、登录会话和 scope 的短期二次验证 Proof"
      },
      "Combination222": {
        "group": [
          {
            "id": 573667
          }
        ],
        "type": "combination"
      },
      "Combination1122": {
        "group": [
...
  },
  "servers": [],
  "security": [
    {
      "Combination343": []
    },
    {
      "Combination1243": []
    }
  ]
}